+++ src/components/SelectSortBy.tsx
@@ -6,7 +6,7 @@ import {
   } from '@grafana/ui';
   import React, { useState, useEffect } from 'react';
 
-export const SelectSortBy = ({ query, updateQuery, datasource }) => {
+export const SelectSortBy = ({ query, updateQuery, datasource, table }) => {
     let sortDirectionOptions = [
       { label: 'ASC', value: 'ASC', icon: 'arrow-up' },
       { label: 'DESC', value: 'DESC', icon: 'arrow-down' },
@@ -19,7 +19,7 @@ export const SelectSortBy = ({ query, updateQuery, datasource }) => {
       let unmounted = false;
 
       async function getTableColumnOptions() {
-        results = await datasource.snowConnection.getTableColumnOptions(query.tableName?.value);
+        results = await datasource.snowConnection.getTableColumnOptions(table);
         if (!unmounted) {
           if (results.length > 0) {
             console.log('Setting tableColumn options: ', results);
@@ -37,7 +37,7 @@ export const SelectSortBy = ({ query, updateQuery, datasource }) => {
       return () => {
         unmounted = true;
       };
-    }, [datasource.snowConnection, query.tableName, query.sortBy]);
+    }, [datasource.snowConnection, table, query.sortBy]);
 
     return (
       <>